# improve parser performance