--- a/src/pytest_bdd/parser.py
+++ b/src/pytest_bdd/parser.py
@@ -292,6 +292,7 @@ class Scenario:
 class Step:
     type: str
     _name: str
+    _full_name: str | None
     line_number: int
     indent: int
     keyword: str
@@ -311,6 +312,7 @@ class Step:
         self.scenario = None
         self.background = None
         self.lines = []
+        self._full_name = None
 
     def add_line(self, line: str) -> None:
         """Add line to the multiple step.
@@ -318,25 +320,29 @@ class Step:
         :param str line: Line of text - the continuation of the step name.
         """
         self.lines.append(line)
+        self._full_name = None
 
     @property
     def name(self) -> str:
-        multilines_content = textwrap.dedent("\n".join(self.lines)) if self.lines else ""
+        if self._full_name is None:
+            multilines_content = textwrap.dedent("\n".join(self.lines)) if self.lines else ""
 
-        # Remove the multiline quotes, if present.
-        multilines_content = re.sub(
-            pattern=r'^"""\n(?P<content>.*)\n"""$',
-            repl=r"\g<content>",
-            string=multilines_content,
-            flags=re.DOTALL,  # Needed to make the "." match also new lines
-        )
+            # Remove the multiline quotes, if present.
+            multilines_content = re.sub(
+                pattern=r'^"""\n(?P<content>.*)\n"""$',
+                repl=r"\g<content>",
+                string=multilines_content,
+                flags=re.DOTALL,  # Needed to make the "." match also new lines
+            )
 
-        lines = [self._name] + [multilines_content]
-        return "\n".join(lines).strip()
+            lines = [self._name] + [multilines_content]
+            self._full_name = "\n".join(lines).strip()
+        return self._full_name
 
     @name.setter
     def name(self, value: str) -> None:
         self._name = value
+        self._full_name = None
 
     def __str__(self) -> str:
         """Full step name including the type."""