--- a/src/pytest_bdd/parser.py
+++ b/src/pytest_bdd/parser.py
@@ -319,8 +319,7 @@ class Step:
         :param str line: Line of text - the continuation of the step name.
         """
         self.lines.append(line)
-        if "full_name" in self.__dict__:
-            del self.__dict__["full_name"]
+        self._invalidate_full_name_cache()
 
     @cached_property
     def full_name(self) -> str:
@@ -337,6 +336,11 @@ class Step:
         lines = [self._name] + [multilines_content]
         return "\n".join(lines).strip()
 
+    def _invalidate_full_name_cache(self) -> None:
+        """Invalidate the full_name cache."""
+        if "full_name" in self.__dict__:
+            del self.full_name
+
     @property
     def name(self) -> str:
         return self.full_name
@@ -344,8 +348,7 @@ class Step:
     @name.setter
     def name(self, value: str) -> None:
         self._name = value
-        if "full_name" in self.__dict__:
-            del self.__dict__["full_name"]
+        self._invalidate_full_name_cache()
 
     def __str__(self) -> str:
         """Full step name including the type."""