Hi,

Using py-spy to check where to improve the performance of my tests, I noticed a lot of time was spent in the Step name property.
Just by caching the value instead of recalculating it every time I gain ~15%.
Same gain when I run the tests in parallel

Sequence:
* Before:
    * 1280 passed, 49 deselected in 125.52s (0:02:05)
* After:
    * 1280 passed, 49 deselected in 107.67s (0:01:47)

Parallel:
* Before:
    * 1280 passed in 60.61s (0:01:00)
* After:
    * 1280 passed in 50.24s

It is a very minor change, hope it can be merged.

Thanks