@dcendents 2023-07-14
Hi @youtux ,

Actually you were right we can `del self.full_name` but we do need to validate the key is in `__dict__` first, which I was not doing the first time around and caused the error I mentionned.

Anyway that showed my modifications were lacking proper testing so I added a unit test as well to validate the cache behavior.
I really tried to use a mock to wrap the step object and spy on the `full_name` method (property) but I could not make it work, so I hope you are ok with the way I tested it.

Changing values directly on the step object to show the cache value is still returned (which should never happen in production code).
Changing values using properties show the cache is invalidated.

Cheers